LayoutTests/platform/chromium-win/http/tests/security/javascriptURL/xss-DENIED-javascript-execution-expected.txt

CONSOLE MESSAGE: Unsafe JavaScript attempt to access frame with URL http://localhost:8000/security/javascriptURL/resources/foreign-domain-iframe.html from frame with URL http://127.0.0.1:8000/security/javascriptURL/xss-DENIED-javascript-execution.html. Domains, protocols and ports must match.

CONSOLE MESSAGE: Unsafe JavaScript attempt to access frame with URL http://localhost:8000/security/javascriptURL/resources/foreign-domain-iframe.html from frame with URL http://127.0.0.1:8000/security/javascriptURL/xss-DENIED-javascript-execution.html. Domains, protocols and ports must match.

CONSOLE MESSAGE: Unsafe JavaScript attempt to access frame with URL http://localhost:8000/security/javascriptURL/resources/foreign-domain-iframe.html from frame with URL http://127.0.0.1:8000/security/javascriptURL/xss-DENIED-javascript-execution.html. Domains, protocols and ports must match.

CONSOLE MESSAGE: Unsafe JavaScript attempt to access frame with URL http://localhost:8000/security/javascriptURL/resources/foreign-domain-iframe.html from frame with URL http://127.0.0.1:8000/security/javascriptURL/xss-DENIED-javascript-execution.html. Domains, protocols and ports must match.

CONSOLE MESSAGE: Unsafe JavaScript attempt to access frame with URL http://localhost:8000/security/javascriptURL/resources/foreign-domain-iframe.html from frame with URL http://127.0.0.1:8000/security/javascriptURL/xss-DENIED-javascript-execution.html. Domains, protocols and ports must match.

CONSOLE MESSAGE: Unsafe JavaScript attempt to access frame with URL http://localhost:8000/security/javascriptURL/resources/foreign-domain-iframe.html from frame with URL http://127.0.0.1:8000/security/javascriptURL/xss-DENIED-javascript-execution.html. Domains, protocols and ports must match.

CONSOLE MESSAGE: Unsafe JavaScript attempt to access frame with URL http://localhost:8000/security/javascriptURL/resources/foreign-domain-iframe.html from frame with URL http://127.0.0.1:8000/security/javascriptURL/xss-DENIED-javascript-execution.html. Domains, protocols and ports must match.

CONSOLE MESSAGE: Unsafe JavaScript attempt to access frame with URL http://localhost:8000/security/javascriptURL/resources/foreign-domain-iframe.html from frame with URL http://127.0.0.1:8000/security/javascriptURL/xss-DENIED-javascript-execution.html. Domains, protocols and ports must match.

CONSOLE MESSAGE: Unsafe JavaScript attempt to access frame with URL http://localhost:8000/security/javascriptURL/resources/foreign-domain-iframe.html from frame with URL http://127.0.0.1:8000/security/javascriptURL/xss-DENIED-javascript-execution.html. Domains, protocols and ports must match.

CONSOLE MESSAGE: Unsafe JavaScript attempt to access frame with URL http://localhost:8000/security/javascriptURL/resources/foreign-domain-iframe.html from frame with URL http://127.0.0.1:8000/security/javascriptURL/xss-DENIED-javascript-execution.html. Domains, protocols and ports must match.

If you see alert windows, the test failed. 


--------
Frame: 'DUPA'
--------
You should not see this in an alert window.
